requirements.txt
gpxpy
paho-mqtt